Implement own shouldComponentUpdate for KendoComboBox
We opt KendoComboBox out of ImmutableOptimizations because it's too
pessimistic; even when passed to _.isEqual, two different function
objects that have the same implementations are not considered equal, so
it's returning true too often.

The reason why we couldn't just make this new shouldComponentUpdate the
implementation for ImmutableOptimizations, is that for different
components, the subsets of props that have value equality semantics and
props that have reference equality semantics is different.